AR

Description

Class for Appearance Records

Documentation

class AR ar where

class for Appearance Records

Methods

arstate :: ar q -> q

access state component of the AR

arstart :: q -> Int -> ar q

initial state to start construction

ardelta :: Ord q => DRA q sigma l -> ar q -> [(sigma, ar q)]

transition function (labelled successors to an AR)

arprio :: Ord q => [Rabinpair q] -> ar q -> Int

assign priorities to ARs

arempty :: q -> ar q

default AR for states in edgeless SCCs or SCCs where no indices have to be tracked

ufix

Arguments

:: Eq a

=> (a -> a)
function

-> a
start

-> a
fixpoint

apply a function recursively until a fixpoint is reached

reachl

Arguments

:: Ord q

=> (q -> Set q)
successor function

-> Set q
set of elements

-> Set q
direct successors of the elements

Computes from a set of elements the set of direct successors.

reach

Arguments

:: (Eq q, Ord q)

=> (q -> Set q)
successor function

-> q
starting element

-> Set q
reachable elements

Computes all reachable elements starting from a given element given a function that computes the direct successors of an element.

prios :: (Ord (ar q), Ord q, AR ar) => [Rabinpair q] -> [ar q] -> Map (ar q) Int

compute priorities to given states and reduce them.

If only priorities 5,6,7 occur they are reduced to 1,2,3.

rabin2parity

Arguments

:: (Ord (rec q), Ord q, AR rec)

=> DRA q sigma l
the Rabin automaton

-> DPA (rec q) sigma Int
the parity automaton having ARs as states

translate a Rabin automaton into a parity automaton

exdra :: DRA Int Char ()

example Rabin automaton (from Example 5.6 of the thesis)

 exdra = makeAutomat
    [ (1,()),(2,()),(3,())
    , (4,()),(5,()),(6,())
    , (7,()),(8,())
    ]
    [ (1,a,2), (1,b,5)
    , (2,a,3), (2,b,4)
    , (3,a,7), (3,b,3)
    , (4,a,5), (4,b,6)
    , (5,a,4), (5,a,5)
    , (6,a,6), (6,b,5)
    , (7,a,8), (7,b,7)
    , (8,a,7), (8,b,8)
    ]
    [1]
    [ (S.fromList [6], S.fromList [5])
    , (S.fromList [4,5], S.fromList [6])
    , (S.fromList [1,7], S.fromList [2,3])
    , (S.fromList [8], S.fromList [2])
    ]