IAR

Description

Translation of Rabin automata to parity automata using Index Appearance Records inspired by https://www.react.uni-saarland.de/teaching/automata-games-verification-11/downloads/ps9.pdf

Documentation

data IAR q

Data structure for Index Appearance Record

Constructors

IAR

Fields

state :: q

the current state

perm :: [Int]

permutation of indices of pairs

deltaIAR :: Ord q => DRA q sigma l -> IAR q -> [(sigma, IAR q)]

transition function of the parity automaton

(q, perm) -> (q', perm') where:

q -> q' : sucessor as in Rabin automaton

second component: sort all indices in perm where q is in the excluded set and move those to the front

prio :: Ord q => [Rabinpair q] -> IAR q -> Int

priority to states in the parity automaton

prio (q,perm): 2*rpos oder 2*rpos + 1