IAR_optimized

Description

Improvements for IAR procedures: SCC decomposition, finding a good initial state.

Documentation

rabin2parity

Arguments

:: (Eq sigma, Ord q, AR ar, Ord (ar q))

=> (DRA q sigma l -> DPA (ar q) sigma Int)
basic translation of Rabin automata to parity automata (e.g. rabin2parity from AR)

-> DRA q sigma l

-> DPA (ar q) sigma Int

translate a Rabin automaton into a parity automaton applying the improvements (SCC decomposition, finding a good initial state in each SCC) using the given basic translation

>>> IAR_optimized.rabin2parity AR.rabin2parity  exdra :: DPA (IAR Int) Char Int

Helper

applytoboth :: (a -> b) -> (a, a) -> (b, b)

apply a function to both components of a pair

>>> applytoboth (5+) (1,5)
(6,10)

fromSubGraph

Arguments

:: Ord q

=> Automat q s l acc

-> Gr l s
the graph of the new automaton (subgraph of the old automaton)

-> Set q
new intial states

-> acc
new acceptance condition

-> Automat q s l acc

compute a new automaton from a subgraph of an automaton

>>> fromSubGraph exdra
        (G.mkGraph [(4,()),(5,()),(6,())] [(4,5,'a'),(4,6,'b'),(5,4,'a'),(5,5,'a'),(6,5,'b'),(6,6,'a')]) 
        (S.fromList [4]) 
        [(S.fromList [6], S.fromList [5]),(S.fromList [4,5], S.fromList [6])]
Automat { states = fromList [4,5,6]
        , bimap = fromList [(4,4),(5,5),(6,6)]
        , graph = mkGraph [(4,()),(5,()),(6,())] [(4,5,'a'),(4,6,'b'),(5,4,'a'),(5,5,'a'),(6,5,'b'),(6,6,'a')]
        , start = fromList [4]
        , accept = [(fromList [6],fromList [5]),(fromList [4,5],fromList [6])]
        , aps = Nothing
        }

larWithState :: (Eq q, AR ar) => [DPA (ar q) s l] -> q -> ar q

find an Appearance Record in the given automata with a given state component

>>> larWithState [IAR_optimized.rabin2parity AR.rabin2parity  exdra :: DPA (IAR Int) Char Int] 4
IAR {state = 4, perm = [1,0]}

sccNodes :: Gr l s -> ([[Node]], [[Node]])

compute SCCs of a graph (as list of nodes) grouped as (edgeless, cyclic)

>>> sccNodes $ graph exdra
([[1],[2]],[[4,5,6],[7,8],[3]])

sccs :: Gr l s -> ([Gr l s], [Gr l s])

compute SCCs of a graph (as subgraphs) grouped as (edgeless, cyclic)

>>> sccs $ graph exdra
( [mkGraph [(1,())] [],mkGraph [(2,())] []]
, [ mkGraph [(4,()),(5,()),(6,())] [(4,5,'a'),(4,6,'b'),(5,4,'a'),(5,5,'a'),(6,5,'b'),(6,6,'a')]
  , mkGraph [(7,()),(8,())] [(7,7,'b'),(7,8,'a'),(8,7,'a'),(8,8,'b')]
  , mkGraph [(3,())] [(3,3,'b')]
  ]
)

interestingsccs

Arguments

:: Gr l s
the graph

-> [([Node], [Node])]
the Rabin pairs translated to nodes in the graph

-> ([Gr l s], [(Gr l s, [([Node], [Node])])])

compute SCCs of a graph and the pairs that have to be tracked in this SCC grouped as (SCCs where no pair has to be tracked, SCCs and pairs that have to be tracked)

>>> interestingsccs (graph exdra) [([6],[5]),([4,5],[6]),([1,7],[2,3]),([8],[2])]
([ mkGraph [(1,())] []
 , mkGraph [(2,())] []
 , mkGraph [(7,()),(8,())] [(7,7,'b'),(7,8,'a'),(8,7,'a'),(8,8,'b')]
 ]
,[ (mkGraph [(4,()),(5,()),(6,())] [(4,5,'a'),(4,6,'b'),(5,4,'a'),(5,5,'a'),(6,5,'b'),(6,6,'a')]
   ,[([6],[5]),([4,5],[6])]
   )
 , (mkGraph [(3,())] [(3,3,'b')]
   ,[([1,7],[2,3])]
   )
 ]
)

interestingfs :: [([Node], [Node])] -> Gr l s -> [Int]

compute indices of pairs that have to be tracked in the graph (the final set intersects the graph)

Example: Let g be the SCC [4,5,6] of the Rabin automaton exdra

>>> interestingfs [([6],[5]),([4,5],[6]),([1,7],[2,3]),([8],[2])] g
[0,1]

btwsccs :: Ord q => DRA q s l -> (q, q) -> Bool

are the two states in different SCCs of the graph of the automaton

>>> btwsccs exdra (4,7)
True

>>> btwsccs exdra (8,7)
False

SCC Decomposition

components :: Ord q => DRA q s l -> ([DRA q s l], [DRA q s l])

decompose and restrict a Rabin automaton to its strongly connected components they are grouped as (edgeless SCC or SCC where no pair has to be tracked, restriction of the Rabin automaton to a component)

connectcomponents :: (Ord q, AR ar, Ord (ar q)) => DRA q s ll -> [DPA (ar q) s Int] -> DPA (ar q) s Int

connect the parity automata of the components

Takes all automata and adds edges between them as in the Rabin automaton.

combineDPAs :: Ord q => q -> [DPA q s l] -> DPA q s l

combine parity automata to one automaton having the automata as components (no new edges are added) and the union of the acceptance conditions

parsuccs :: (Eq q, Ord q, AR ar, Ord (ar q)) => DRA q s ll -> [DPA (ar q) s l] -> ar q -> [(ar q, s, ar q)]

compute the successors of an Appearance Record in different SCCs

Example: Let dpas be the parity automata of the SCCs of the Rabin automaton exdra

>>> parsuccs exdra dpas (IAR 3 [0])
[(IAR {state = 3, perm = [0]},'a',IAR {state = 7, perm = []})]

toDPA :: (Ord q, AR ar, Ord (ar q)) => DRA q s l -> DPA (ar q) s Int

direct translation of Rabin automata with empty acceptance condition or no edges

Example: Let tdra be the Rabin automaton exdra restricted to its component [7,8]

>>> toDPA tdra :: DPA (IAR Int) Char Int
Automat { states = fromList [ IAR {state = 7, perm = []}, IAR {state = 8, perm = []}]
        , bimap = fromList [(IAR {state = 7, perm = []},7),(IAR {state = 8, perm = []},8)]
        , graph = mkGraph [(7,1),(8,1)] [(7,7,'b'),(7,8,'a'),(8,7,'a'),(8,8,'b')]
        , start = fromList [IAR {state = 7, perm = []}]
        , accept = MaxEv {priomap = fromList [(IAR {state = 7, perm = []},1),(IAR {state = 8, perm = []},1)]}
        , aps = Nothing
        }

Initial State

reduceparitygraph :: (Ord q, AR ar, Ord (ar q)) => DRA q s ll -> DPA (ar q) s l -> DPA (ar q) s l

pick a good initial state by choosing the BSCC of the parity automaton

Example: let tdra be the Rabin automaton exdra restricted to its SCC [4,5,6]

>>> reduceparitygraph tdra $ AR.rabin2parity tdra :: DPA (IAR Int) Char Int
Automat { states = fromList [ IAR {state = 4, perm = [1,0]}
                            , IAR {state = 5, perm = [0,1]}
                            , IAR {state = 5, perm = [1,0]}
                            , IAR {state = 6, perm = [0,1]}
                            , IAR {state = 6, perm = [1,0]}
                            ]
        , bimap = fromList [ (IAR {state = 4, perm = [1,0]},2)
                           , (IAR {state = 5, perm = [0,1]},3)
                           , (IAR {state = 5, perm = [1,0]},4)
                           , (IAR {state = 6, perm = [0,1]},5)
                           , (IAR {state = 6, perm = [1,0]},6)
                           ]
        , graph = mkGraph [(2,1),(3,3),(4,2),(5,2),(6,3)] 
                          [(2,4,'a'),(2,6,'b')
                          ,(3,2,'a'),(3,4,'a')
                          ,(4,2,'a'),(4,4,'a')
                          ,(5,3,'b'),(5,5,'a')
                          ,(6,3,'b'),(6,5,'a')
                          ]
        , start = fromList [IAR {state = 4, perm = [0,1]}]
        , accept = MaxEv {priomap = fromList 
                [ (IAR {state = 4, perm = [0,1]},3)
                , (IAR {state = 4, perm = [1,0]},1)
                , (IAR {state = 5, perm = [0,1]},3)
                , (IAR {state = 5, perm = [1,0]},2)
                , (IAR {state = 6, perm = [0,1]},2)
                , (IAR {state = 6, perm = [1,0]},3)
                ]
                         }
        , aps = Nothing
        }